requirements.txt
pyyaml
kubernetes
jinja2
